requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
feedparser==6.0.11
requests==2.32.3
tenacity==9.0.0
pydub==0.25.1
moviepy==1.0.3
Pillow==10.4.0
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
tqdm==4.66.4
